#458 added overlay and ESC key functionality
