+++ packages/popover/src/Popover/index.js
@@ -1,2 +1,2 @@
-import Popover from './Popover'
+import Popover from './PopoverController'
 export default Popover
